[forwarder] allow multiple endpoints
In order to do so, the config was modified:
other_dd_urls and other_api_keys were added.
They both accept a a comma-separated "list", which is
transformed in a endpoints dict. (merged with dd_url and api_key)
This one has the following format:
{
  'https://app.datadoghq.com': ['api_key_abc', 'api_key_def'],
  'https://app.example.com': ['api_key_xyz']
}

The forwarder is the only one using this new configuration.
It loops through all endpoints and sends the same payload.

[tests] multiple endpoints logic

Config tests and update transaction tests.
Set default max_parallelism to 1 in transaction.py. (it's the real
default)